Ensure consistent short sha in perfetto ui zip
Specify value of 14 when getting the short commit sha to use
in the perfetto ui zip. This ensures we get the same length
for ay given commit, regardless of which version of git
is being used. In 2016, git was updated to use a default
short sha length based on the size of the repo. That means
that --short can return a different length depending on the
git version. 14 should be more than plenty to ensure
uniqueness