ui: Make the overview timeline fast again
The overview timeline for sched breaks the timeline in 100
steps and then makes a query for each 1% of the timeline.
The previous code structure was kicking off the sql requests
in sequence, one per each JS microtask, only after the previous one
was completed. Unfortunately this would cause the browser to
interleave each of the 100 microtasks with animation frames if the
microtasks were long enough. This means getting a scheduling like
[Step 1][Animation frame][Step 2][Animation frame][Step 3]...
If the animation frame itself takes 8-16 ms (because we render a
lot of stuff), this introduces an artificial delay of 8-16 ms between
each step (or between each few steps), which is unideal.
This change instead queues the 100 microtasks upfront. They are still
executed in sequence, but because they are queued together, they take
precedence over other animation frames that happen later.

Change-Id: I460df007a7dbb0625a91844fa1cbd71489d006cb