trace_writer_impl: Avoid using Message::inc_size_already_written
Previously, TraceWriterImpl had two support two different ways to write
a TracePacket fragment size into the chunk:

1. When the C++ protozero::Message was used:
  1. `TraceWriterImpl::NewTracePacket()` returned a protozero::Message
    that had `size_field_` pointing to the fragment size.
  2. When a new chunk was created:
    * The old fragment size would be filled by TraceWriterImpl.
    * `size_field_` would be redirected to point to the new fragment
      size and inc_size_already_written() would be used.
  3. When `protozero::Message::Finalize()` was be called on the root
     TracePacket message, it would fill the last fragment size.
2. When the protozero::MessageHandle::TakeStreamWriter() was used
   (mostly from the C shared library):
  1. `TraceWriterImpl::NewTracePacket()` was used to extract a
     ScatteredStreamWriter.
  2. When a new chunk was created:
    * The old fragment size would be filled by TraceWriterImpl.
  3. When the user finished writing the packet, they would call
     `TraceWriterImpl::FinishTracePacket()` and TraceWriterImpl would
     fill the size for the last fragment.

This commit is based on the observation that the write in 1.3 above is
useless. Whoever is scraping the chunk (the tracing service) will never
read the size of the last fragment until the fragment is marked as complete.

With that observation we can unify the code paths:
inc_size_already_written is not used anymore. Instead, TraceWriterImpl
will fill the correct size of the fragment on FinishTracePacket (if
that's called at all) or on the next NewTracePacket or Flush.

Bug: 162206162
Change-Id: Idae4cff8674fbeb67226212f6de85703c1ba9356