trace_writer_impl: Keep spare packet to ease scraping
The tracing service scraping logic ignores the last packet in the chunk
when a chunk is not complete because it doesn't know when a trace writer
is done with a packet.

This means that, without additional workarounds, the tracing service is
often not able to get the last TracePacket for data sources that do not
properly handle flush (such as TrackEvent).

This commit implements a workaround in the client (TraceWriter): when a
data source marks a packet as complete (FinishTracePacket), it creates
a spare packet, which will not be visible to the tracing service, but it
will make the previos visible.

* If the user calls NewTracePacket(), TraceWriterImpl can return the
  spare packet, if it has one.
* If the Flush() or GetNewBuffer() is called, the spare packet will be
  finalized and made visible.

Note that FinishTracePacket() is used only by the shared library SDK
data sources. This commit doesn't fix the problem for C++ SDK data
sources (not even TrackEvent) or C++ producers. Future commit might fix
those (by explicitly including FinishTracePacket calls, by changing
the MessageHandle destructor to issue a FinishTracePacket call, or by
providing a new RAII wrapper).

Bug: 162206162
Change-Id: I1cd8dac2eb388fabc061584de36cf0f3064c6df7