tp: fix CREATE OR REPLACE PERFETTO FUNCTION in RUN_METRIC
SQLite does not like functions being dropped in the middle of a function
execution. This is a problem for the implementation of CREATE OR REPLACE
PERFETTO FUNCTION which relies on this behaviour: this breaks inside
RUN_METRIC.

Fix this by not unregistering/registering the function for CREATE OR
REPLACE instead just updating the backing state. While doing this, I
realised all the metrics had broken usage of CREATE PERFETTO FUNCTION
instead of OR REPLACE variant so fix this as well.

Change-Id: I156ace369188026ae8c074a3f98f25eacea067b7