[ui] Fix `registerStaticTrack()`
`registerStaticTrack()` should both register the track and add it to the
default track list - i.e. the equivalent of doing `registerTrack()` and
`addDefaultTrack()`.

Previously, `registerStaticTrack()` did the same thing as
`registerTrack()` as the part where it was supposed to add the track to
the default track list was commented out.

This patch fixes this and makes `registerStaticTrack()` work how it's
supposed to.

The fallout is that every track which is added in `TrackDecider`
needed to be switched from using `registerStaticTrack()` to just
`registerTrack()` to avoid the track showing up twice.

Change-Id: Ib60fa93ae94ea59662b2583a29c2992254a37741