[ui] Add section about tracks to plugin documentation
This CL adds documentation describing how to register and manage
tracks from a plugin.

It purposely avoids going into detail about how to implement custom
tracks as this API is unstable and likely to change in the near
future. The main focus is on the registration of tracks and how to
add them to the timeline.

Change-Id: I21cd39f08275167dd8817e9ee40f5dd82f5e91d8