[stdlib]: Add ts and dur proxy for async binder txns
Async binder txns have dur=0 because the client doesn't wait
for a response. In practice, we can often see the aidl name
slice in a trace which tells us *at least* how much time the
async txn consumed on the server side.

We expose this aidl slice as aidl_ts and aidl_dur:

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*binder.*'
Change-Id: I84fca42eb2edd0dfaf4dc1ad07d67303bb86bc79