tprobes: cpu_reader: continue parsing instead of crashing on ring buffer layout errors
Previously, traced_probes immediately crashed if a raw tracing page did
not match our understanding of the ring buffer ABI. For example when the
iterating the event headers did not stop at the same offset as the valid
data length indicated by the page header. The crashes were a clear and
explicit way of highlighting parsing errors, but we are now revising
this decision.

This patch makes ring buffer ABI errors non-fatal. We still discard the
remainder of the offending page, but the parsing continues from the next
tracing page (and for the rest of the tracing session). Additionally,
the errors are now reported in the trace itself.

I am removing the ZeroPaddedPageTail workaround, with the gotcha
that we won't be able to distinguish where the zero-sized invalid data
record occured in the page. Specifically, we won't know if it's the
previously seen trailing zeroes from buggy(?) kernels, or a completely
unexpected zero-sized record in the middle of a page. This seems
acceptable as I'm not entirely sure the ZeroPaddedPageTail approach was
entirely correct in cases where the kernel stashes lost event counts at
the end of the page (RB_MISSED_STORED).

I am also removing the logging of the offending page, as this has not
proven useful in tracking down the rare errors. We at most get android
tombstones, which only retain several loglines worth of context.

Bug: 204564312
Change-Id: I73a0a37fa173e802a951c12a8b7c85aee28e398f