Add blocked function for systrace displayed in Perfetto UI
When loading a Systrace file in the Perfetto UI, the "Blocked Function" information is not displayed in the details panel when clicking on a thread_state slice in the "Uninterruptible Sleep" state. This information can be readily obtained by parsing a Systrace line with the event_name "sched_blocked_reason", and then presenting this crucial information for analysis in the Perfetto UI. The enhancement not only facilitates visual representation but also supports SQL queries.

There should not be a safety concern, as the Systrace file is human-readable text, for example:"sched_blocked_reason: pid=14829 iowait=1 caller=folio_wait_bit_common+0x294/0x3b0" which indicates that users can directly read the function using a text editor before it is displayed in the Perfetto UI. However, interpreting these details through an editor is far less intuitive compared to a timeline-based UI.

Change-Id: I0144da6c93fc3b82296f336fc7479c67c69015a3