tp: remove experimental package and diffuse contents into other packages
This CL removes the central "experimental" module from the stdlib. This
module's design was a bit of a mistake as it became a central dumping
ground for randoming things without clear ownership. Also making the
tables visible in the documentation while not being something people
should depend on was a bit of a misnomer.

Instead diffuse the modules inside into other packages and make all the
experimental_ prefixed tables just _ (i.e. internal tables which ~have
the same guarantees of being breakable at any time).

Change-Id: I7f2b85476f6bb53690ec2900d542463f184900d6