Add presubmit to prevent modifying the chrome stdlib
This presubmit fails for changes to the src/trace_processor/perfetto_sql/stdlib/chrome or test/data/chrome directories. This because the source of truth lives in Chromium so any changes should be made there instead.
It allows CLs from the Copybara service, which specifies the COPYBARA_IMPORT flag in the CL description.
Manual rolls can still be done with the CHROME_STDLIB_MANUAL_ROLL flag.

Bug: b/324211825
Change-Id: I8f2d44b4f65e2e79327fa91dd4bdd014ce6ccb17