syscall_nr->name tables: update to v6.7 and simplify ARM naming
Handwavy: AArch32 is an execution state of a 64 bit arm processor when
it is running a 32 bit process (backwards compatible with old arm/thumb
machine code). For such a process, the linux syscall table is the same
as that for the older 32 bit armv7 ISAs. So we can drop one of the
tables.

Also, simplify the naming of the tables since the current names are
mixing different terminologies. Go with labels that humans understand,
even if they're not pedantically correct.

Bloaty report for stripped libperfetto.so built with clang-release on
x86_64:
    FILE SIZE        VM SIZE
 --------------  --------------
  +1.0%    +800  +1.0%    +800    [section .rodata]
  +300%     +24  +300%     +24    [LOAD #1 [R]]
  -0.0%     -16  -0.0%     -16    [section .data.rel.ro]
  -0.0%     -16  -0.0%     -16    [section .eh_frame]
  -0.0%     -24  -0.0%     -24    [section .rela.dyn]
  -0.0%     -32  -0.0%     -32    [section .text]
  +0.0%    +736  +0.0%    +736    TOTAL


Change-Id: Iab947b1a6455fd0c69ef30a52807b6945201e7fc