Add perfetto persistent tracing
If a property is set (by shell or by a device build file), perfetto will
keep a simple tracing configuration active all the time.

One test needs to be updated because it assumed that no session with
bugreport_score was running on device. If the property is enabled, that
might not be the case anymore.

Bug: 325622427
Change-Id: I7663f121b44c67ea8d8e23a88c6fb1b588100bfa