perfetto_cmd: show the "No PTY" dialog only when TERM is populated
The "warning: No PTY" dialog is only useful to tell people that
"adb shell perfetto" by itself won't gracefully stop on Ctrl+C
without a -t arg.
However, I noticed that this warning gets printed a lot in tests
as well (when we spawn perfetto as a subprocess).
This CL conditions the warning by the presence of a TERM variable,
which is set by adb shell but is usually not set in test environments.
It's not a perfect solution, but strictly better than the past.

Change-Id: I7dcba60f7fee0ac2ad64e424ee9fba477a40c985