GN Standalone: use .rsp files
Response files avoid building up uber-long cmdlines by packing
all inputs in one file and passing the file on the cmdline.
We were not using rsp files on solink and link, but we were using
it only for alink (for static libraries)
This change makes use of .rsp files on "solink" and "link" when
linking .so and executables.

Test: manual: builds on both MacOS and Linux Debian
Bug: https://github.com/google/perfetto/issues/718
Change-Id: Ie660a181a999728ca99f273929af2d7be783f1aa