Fix TraceBlob move assignment operator
There was a subtle bug in TraceBlob move assignment operator, where the
memory owned by this trace blob would be leaked.

Fix this, as suggested by Primiano, by explicitly implementing the move
constructor and using the move constructor from the move assignment
operator.

This was not an actual problem, nothing in the codebase calls TraceBlob.
I'm fixing it anyway because TraceBlob is part of the public interface.

Change-Id: Ia0bed4c89612470abc80ccbfd0df9ea3434c180b