ftrace: default to 8MB per-cpu buffer_size_kb on devices with 8GB+ of ram
The config defaults should appropriate for most users, and we can be
more lenient with buffer sizing on devices with sufficient physical
memory. Lower memory devices will keep getting the historical default of
2 MB per cpu.

While we can change the default for buffer_size_kb, we should also
continue honoring the existing configs. That, and having an option to
force an exact size is useful by itself.

To help with configs that need to run on both older and newer perfetto
builds (while taking advantage of bigger defaults based on ram size),
I've introduced buffer_size_lower_bound. See its proto comment for a
description, but the short version is:
* new local configs should set neither buffer_size_kb or
  buffer_size_lower_bound -> we pick a good default
* cross-build configs should set both fields, and the resulting buffer
  will be no less than requested.
* for specific needs, buffer_size_kb by itself remains an exact value
  (+- guardrails for unrealistic values).

Note: we actually check for 7GB or above since sysconf doesn't return a
power of two. I didn't investigate whether this is to do with
hw-reserved memory or other factors. A test phone with 8GB quoted mem
reported ~7.2GB through sysconf.

Change-Id: I011ac75795e48f895228da6375b81a361a99e05e