[stdlib]: Add waker_id to the thread_state table
The thread_state table has a waker_utid column but what's often
needed in the end is the thread_state id in the waker thread.
Getting this requires an expensive timestamp BETWEEN JOIN.

We now add the waker_id directly in the table, this gives us the thread
state id of the Running thread_state that woke the thread up and we are
just a LAG() away from the Runnable state.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell
Change-Id: Iee269f05c2d59713433f2dea8fd9b0f8e3948d04