tp: make QueryExecutor benchmarks more meaningful
This CL improves the QueryExecutor benchmarks to make them actually
return data for most of the queries. If we return no rows, that's the
best possible case as much of the query codepaths become very optimized.
This is also by far the less common case as most of the time, when users
query some data, it's because it actually exists. Therefore we should
optimize for this case.

Change-Id: I15e3413cea1b3d15cd054527109f4ff04d9ce41b