tp: make SetBitsIterator an implementation detail of BitVector
This CL makes the SetBitsIterator a private implementation detail of
BitVector by introducing two new methods: GetSetBitIndices and
SelectBits. These two methods are the last ways in which the
SetBitsIterator was being used.

Follow-up changes will change the implementation of these methods to
be more optimized allowing SetBitsIterator to be removed entirely.

Change-Id: I158b3c289a642f8c6647cccd2f6d676c64ba821b