tp: parse new procfs info into process.start_ts and new runtime counter tracks
There are two new pieces of info being serialised in the trace:
* process descriptions (process_tree) include the bootclock timestamp of
  when a process was started.
* polled process counters (process_stats) include the total-so-far of
  time scheduled in userspace and kernel mode.

The latter is straightforward, we simply add new process-scoped counter
tracks.

The earlier is a little bit more interesting in that we already have
"process.start_ts", but it's only populated if we observe a process
being spawned during a trace. This is based on ftrace events with
nanosecond resolution, while the new procfs timestamps I'm adding in are
at 10ms resolution. Users might find it surprising that the same column
has potentially different precisions, but I've decided that it's still
sensible to share the same column.

See r.android.com/2982111 for the recording side.

Bug: 326892115
Change-Id: If109c1fc3c595214533f98efd799fa6d9c1f70cd