Support build_all_channels for release branches
builder_entrypoint.sh is executed on every:
- UI / TP commit
- Every change to a release branch
(as configured via gcloud)

However prior to this change the SHA of the *triggering commit*
was ignored. Instead we:
- unconditionally checked out the Perfetto repo
- looked at ui/release/channels.json
- checked out the three revisions detailed there
- for each revision we build and uploaded the UI
- for stable we additionally update index.html & the service worker.

After this change we add support to builder_entrypoint.sh and
build_all_channels.py for building just a specific commit and uploading
that. We use this for release branches which we detect this by
examining the triggering commit and seeing if the commit is the
HEAD of a branch matching origin/releases/v*.

Change-Id: I6705816de85fcf60a22afdbe3243e71875b17523