Reapply "Add MappingTracker"
This reverts commit e5df54a90ec655ec4c47b7d26b0de7a54c98d75f.

The reason the tests were failing is that the new code is more strict in
determining whether a build_id is a hex module id. The old code just
uses a length of 33 to determine an id is a module_id. My new code was
also checking that the build_id was a hex string. Turns out some tests
just use dummy 33 length strings.

So I removed the extra checks (see diff patchset 1 vs 2)

Bug: b/283794416
Change-Id: I6ea5653534a258c7ca25b9bd96df56a107bcddd3