tp: simplify IndexSearch's API
This CL simplifies the API and usage of DataLayerChain::IndexSearch.

Specifically, the purpose of this CL is to remove the two-step process
of a) creating a range or BitVector to update the indices b) applying
the range/bitvector to update the index vector.

Instead of this, implementations are now expected to directly remove
indices. The "payload" keeps the original index while the "index" keeps
being modified as we move down the chain.

The reason for this CL is to remove expensive BitVector reconcile
operations which are totally unnecessary: we can just have a downword
facing translation instead of both downward and upwards translation.
It also makes the code a lot simpler to reason about.

Change-Id: I1b04aef7ceba32dcdba209a3be8132ebc6fd5f1d