Rework timeline subset range as an async callback function
It turns out that Sokatoa, when hosting Perfetto, needs to calculate
frame-based timestamp ranges to constrain the Perfetto timeline from
the slices in the Perfetto trace, itself. So they cannot be computed
before opening the Perfetto app and it is necessary for Perfetto to
call back into the host application, which then uses the Engine to
query the trace database.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>