Fix build with LTO and new LLVM
Instantiating the InProcessTracingBackend vtable requires instantiating
the destructor. Since InProcessTracingBackend owns a
unique_ptr<TracingService>, the destructor needs to see the full
TracingService definition, not just a forward declaration.

So far, everything was working fine. It appears that newer versions of
LLVM with LTO need to instantiate the vtable somewhere else and complain
TracingService need to be visible.

Since we cannot make TracingService visible, this simply moves the
destructor out of line.

Also, user `= default` instead of `{}`, as a matter of style.

Bug: 329245881
Change-Id: I9cad6bb9a2b97a8c729a0d65cd13ea2c6451d374