tp: Implement interval intersect prototype
C++ based implementation of interval intersect (replacement for span join) without partitioning, so it can only be exposed as internal (prefixed with _) macro.

Adds a new StaticTableFunction and an SQL macro that is responsible for invoking it. The current complexity is O(S1+S2), where S1 is the size of the first interval table and S2 the size of the second. Can be improved to O(S1 + log(S2)) with binary search, but it's not a priority now.

Change-Id: I19c431abceb9598ba94db6b049a61d671964f524