Speed up HeapprofdJavaCtsTest Oom tests
The Oom workflow for the perfetto java plugin works as follows:

1. A user configures a tracing session with a start trigger.
2. The app OOMs.
3. The app (after forking) connects to perfetto and registers a data
   source.
4. The app sends a trigger.
5. perfetto sees the trigger, starts the configured tracing session and
   starts the app data source.
6. The app waits 1s for the data source to start.
7. The app receives a start notification and starts writing data.
8. perfetto stops the tracing session after `stop_delay_ms` and sends a
   stop command to the app data source.
9. The app finishes writing data.
10. The app replies to the stop command.
11. perfetto waits for the stop acknowledgement from the app. perfetto
    will wait up to `data_source_stop_timeout_ms`.

In the test, `stop_delay_ms` is very long (10s). So, most of the times,
9 will happen before 8 and the test is slower than necessary.

`stop_delay_ms` can be shorter, because, once the data source has been
started, perfetto will wait for the app no matter what.

This appears to speed up the test from ~11s to ~3s on an x86_64
emulator.

Change-Id: I159b0f528bb9a1ba4ade6db97a6b1968f30af1c8