Explicitly install optional components of perfetto
perfetto-extras is a phony module that abstracts optional components for
perfetto. Previously, the dependencies were installed by directly
required from traced_probes.

But that is expected to interfere with an upcoming change in Soong where
`required` dependencies are installed by Soong. Specifically, it will
increase the size of the `microdroid` module by 2-3MB because the
dependencies - which are unnecessary in Microdroid - will be installed
there due to the `required` being effective in Soong.

This change fixes that regression by making the dependencies to be only
for Android.

Bug: 321626681
Test: check the dependencies are found in Android
Test: check the dependencies are absent in Microdroid
Change-Id: I01de13d6a65c683302f85fad69107171c1f2022e