Add CPU filter into traced_perf
PMUs can generate large amounts of data which causes an overhead in term of
capture and processing. Perfetto capture the same PMUs for all of the CPUs.

This change introduce parameters to restrict capture on either specific CPU
of the system or CPUIDs. The capture can be configured to restrict data
captured to CPUs that actually matters.

It also introduces a parameter to gracefully fail if a perf event cannot be
open; for example in the case it is not available or supported on a given CPU.