shared_lib: Categories in track_event are disabled by default
When instrumenting the code with the C track_event API, like:

```
  #define EXAMPLE_CATEGORIES(C)                                   \
    C(rendering, "rendering", "Rendering events")

  PERFETTO_TE_CATEGORIES_DEFINE(EXAMPLE_CATEGORIES)

  void foo() {
    PERFETTO_TE(rendering, PERFETTO_TE_INSTANT("rendering_event"));
  }
```

Users are surprised that a configuration like the following:

```
data_sources {
  config {
    name: "track_event"
    track_event_config {
    }
  }
}
```

Enables the "rendering" category even if it's not explicitly listed.

Since the C API is relatively new and doesn't have a lot of users yet,
change the default behavior: categories are disabled by default.

In order to enable the all categories in a config, users can have:

```
  config {
    name: "track_event"
    track_event_config {
      enabled_categories: "*"
    }
  }
```

Change-Id: I2104cf1da2899897c64bd0f6318c9715ae79d234