Trace Redaction - Timeline and Build Timeline
The timeline data structure is critical to future operations. The
timeline is a query optimized structure that, given a pid and ts,
will traverse the thread hierarchy to find which package the
thread belongs to.

The thread-to-package relationship is not fixed. Thread id recycling
means that threads will wonder between packages. This means a thread
that didn't need to be redacted, could suddenly need to be redacted.

Bug: 318576499
Bug: 318576092
Bug: 319850081
Change-Id: Iff0d20d5d762b70898c1f3680af3430ac8262cf1