ftrace: unpessimize last_read_event_ts in case of concurrent data sources
Since we reparse the same batch of pages for each concurrent data
source, this was leaking the "last read" timestamp across them. So the
second DS would end up recording its own last event as the previous
bundle's stopping point.

To repro, start 2 concurrent traces: first with a long duration so that
it is always active, and the second with a RING_BUFFER perfetto
overwrite policy. This will cause all bundles in the second trace to
have incorrect timestamps.

Note: the test changes are unrelated, they would've passed before this
fixup as well. We don't really have tests at the |ReadAndProcessBatch|
level.

Fixes: a9f287ea4 (ftrace: write "last read event timestamp" to assist parsing of ring buffer traces)
Bug: 192586066
Bug: 330438269
Change-Id: I44243de826526dd5166ff784ddd5fc7fa53562f5