tp: multi-machine: insert machine ID to storage tables
This CL retrieves machine_id from TraceProcessorContext for storage
table insertion by trackers (ProcessTracker, TrackTracker, etc.).

The ProcessTracker is updated to use the inserted row number as the
utid/upid of the swapper thread, as the assumption that swapper's
utid/upid is 0 will not hold in injesting a multi-machine trace data
file.

Bug: 284258446

Change-Id: I68e0989df9762dd6d2a4807a75b8b51b73e9f85c