tp: multi-machine: support multi-machine trace sorting
This CL adds the support of sorting multi-machine trace data to the
trace sorter. TraceSorter's Push*Event() methods are extended with a
Machine ID argument that is used to identify trace events from
different machines. The callers pass the TraceProcessorContext machine
ID in pushing the trace events. The shared trace sorter instance will be
used to sort trace data across queues of different machines. In trace
event extraction, the trace sorter parses the trace data using the
parser associated with the machine ID.

Bug: 284258446

Change-Id: I85f3488e9e3b6d3a8fe26cd7518f593f43670457