Actually fix profiling tests for multiuser
51ba1bc15b73("Fix heap profiling tests for multiusers") switched from
reading from "/sdcard/Android/data/" to using a content provider.

But, on multiuser devices, we need to specify `--user` when asking for a
content provider.

The `--user` option and the `cmd user get-main-user` command are only
available from android U. These tests need to run on android S as well
(not in multiuser configurations), so check the SDK version.

Bug: 331275172
Bug: 330278944
Change-Id: If7fbf48c6e5f11f5f9df4d74f920fdbc73ef4801