[stdlib]: Add a floor/ceiling option to the weight bounded DFS
The original wb-DFS ended the reachability search right before a
node would exceed the |root_max_weight|. The critical path algorithm
actually requires the algorithm to add the last node that would exceed
the max weight. This ensures that we follow the critical path right
passed the prev node of the root node.

Exposed both behaviors controlled with an is_target_weight_floor
parameter:
-If true, the root_target_weight is treated as a floor
and the behavior is to include the node that just pushes the
weight above the threshold.
-If false, the root_target_weight is treated as a ceiling
and the behavior is to NOT include the node that would push the
weight above the threshold.

Also updated the argument checking to return early if root_table
or graph_table is empty instead of throwing an error.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*weight_bounded.*'
Change-Id: I035561c3a619961bf3fa5d94a71e926d7b5bdb74