ui: Fix issue where crashed debug tracks cannot be closed
When tracks crash, we want to avoid calling into any of their methods,
including getTrackShellButtons(). This was an issue for debug tracks as
it means their close button does not show up on crashed tracks and thus
they cannot be removed.

This CL:
- Makes the "track close" button part of the core.
- Adds a `closeable` option to the addTracks() action.
- Set this option when adding debug tracks.

Note: Most tracks are not closeable, only specified debug tracks.

Note: This change updates the shape of the state object, but seeing as
the closeable flag is optional, it is forward and backward compatible.

Change-Id: I45bcc62173e028e70de9b93316c60909821a5a8b