Trace Redaction - Create shared filter ftrace events primitive
There are multiple primitives that filter conditional drop ftrace
events. When these are individual primitives, each packet is
deserialized and reserialized N times (where N is the number of
filter primitives).

To reduce the number of reserializations to 1, this CL introduces
a primitive that accepts filters. Each event is tested against
the filters and if any filter say "drop event", the event is
dropped and all later filters are skipped.

Bug: 332547909
Change-Id: I6c9cb665863aafd67d42adc84e3a0fb474d953da