Change 'Remove track' to 'Hide track' (#104)
With the upcoming introduction of user-defined tracks (e.g.,
the user combines one or more counter tracks into a single one),
we need to adjust the existing removal track button. This button
doesn't technically remove the track; it hides it. The track
definition ends up in a filtered list, so that it can be
reintroduced into the timeline by the user via the Track
Filtering timeline toolbar button.

We also need to allow for a track removal that is a "true" removal.
Such an action will do more than just hide the track. It will
delete the track altogether. The user won't be able to un-hide it.
This type of action will be available only to user defined tracks.
The user can of course re-define the track from scratch
to effectively re-introduce it into the timeline.

The existing button is being given a new tooltip and icon to
better reflect the hiding nature. Additionally, the RemoveTrack
action now supports a purge option for use in a true removal.
In that case, we simply don't add the track definition to the
filtered list. We'll use the garbage can icon for that action,
and the tooltip will say "Delete user-defined track". Only
user-defined tracks will have this action, and it will be added
to the track by Sokatoa, as support for these user-defined tracks
is entirely done within Sokatoa using perfetto plugins.