Trace Redaction - Common Redact FTrace Event Modifications
Just like dropping ftrace events, there are/will be multiple
primitives that need to remove individual fields within an event.

This creates a new primitive that allows redactions to be added to
the primitive.

There are a couple limitiations to this. It can only act on the
event within the ftrace event (e.g. sched switch). Only one
redaction can be registered per event (e.g. one redaction for sched
switch).

Change-Id: I80bed18e6bc1ee46016366cdb4904822b288ca15