ftrace: Support atrace_categories_prefer_track_event
A new flag is introduced to the ftrace configuration. It will signal the
framework that for certain atrace categories, the track_event data
source (if available) will be preferred.

If there's one active configuration that doesn't prefer track_event for
a category, the muxer switches back to the legacy implementation.

Bug: 303199244
Change-Id: Ic1164ecb14bcb9b201aea25148f3d8c4cb527757