ui: Remove button backgrounds by default, use intent to add color
Currently, buttons have a blue background by default, but the "minimal"
attribute may be used to turn the background transparent.

It turns out that most buttons in the UI don't want a background, so
have this minimal attribute enabled.

This CL makes minimal the default, and also changes the minimal option
to "intent" instead, which is similar to other UI component libraries.

For now there is only one supported "intent" option "primary", which
turns the background blue, like the default behaviour prior to this
change.

Change-Id: I78634cded282eefbbabb88c915099f518071e128