Trace Redaction - Use Three Stage Pattern In Collect Primitive
The collect primitive was designed to handle cases where
primitive didn't need to search the whole trace. This was only
needed by one primitive. This behaviour introduced more complexity
than value. It is removed in this changed.

Removing the early exit option, introduced a problem with verifying
the context value. To address this, a Begin() and End() optional
functions were added.

Change-Id: Icbf2f8089a98bb5a1faba1530f1db549b1d24dae