add support for multiple constraints in trace URI resolver
Feature request: We'd like the URI provided to the trace resolver to support multiple constraints. For example, the current trace URI will not be able to support URIs like 'a>1;a<3', since keys are stored in a dict, and later values will overwrite the previous one. With this change, we can support such cases, because for constraint classes, we are storing their values as an array, where each element will be a constraint object.

The original discussion in https://github.com/google/perfetto/issues/752

Change-Id: I189af4333409e2dba4b891d85d1d83170ff1135f