ui: Don't blur search bar when hitting backspace
This regression was introduced in the Omnibox refactor in aosp/2679083

Since this CL, pressing backspace in an empty search box would cause it to blur.

This is unintended behaviour and is quite annoying as it means the
search box can lose focus while under normal use.

This is because there is no distinction made between closing the
Onmibox by pressing Esc/clicking outside and hitting backspace when
the text field is empty - all actions call the `onClose` callback.

The reason this feature exists at all is to transition back from query
and command mode back to search mode when the user hits backspace, as
if they were deleting the key character used to transition between
these modes in the first place.

This CL fixes the issue by adding a new callback `onGoBack` on the
Omnibox widget, which is invoked when pressing backspace when the
field is empty.

Command and query modes have been updated to implement this callback
and go back to the search mode, and the search mode ignores it.

Bug: 330781151
Change-Id: I092a347bd50e2fed0675e73c6f79dfc017660f16